Convert CoroFlow actors to standard coroutines
