+++ fdbserver/coroimpl/CoroFlowCoro.cpp
@@ -1,5 +1,5 @@
 /*
- * CoroFlow.actor.cpp
+ * CoroFlowCoro.cpp
  *
  * This source file is part of the FoundationDB open source project
  *
@@ -20,24 +20,20 @@
 
 #include "fdbserver/CoroFlow.h"
 #include "flow/ActorCollection.h"
+#include "Coro.h"
 #include "flow/TDMetric.h"
 #include "fdbrpc/simulator.h"
 #include "fdbrpc/SimulatorProcessInfo.h"
-#include <boost/coroutine2/all.hpp>
-#include <boost/coroutine2/coroutine.hpp>
-#include <functional>
-#include "flow/flow.h"
-#include "flow/network.h"
-#include "flow/actorcompiler.h" // has to be last include
 
-using coro_t = boost::coroutines2::coroutine<Future<Void>>;
+// Old libcoroutine based implementation. Used on Windows until CI has
+// boost context installed
 
-// Coro *current_coro = 0, *main_coro = 0;
-// Coro* swapCoro( Coro* n ) {
-// 	Coro* t = current_coro;
-// 	current_coro = n;
-// 	return t;
-// }
+Coro *current_coro = 0, *main_coro = 0;
+Coro* swapCoro(Coro* n) {
+	Coro* t = current_coro;
+	current_coro = n;
+	return t;
+}
 
 /*struct IThreadlike {
 public:
@@ -51,57 +47,27 @@ protected:
 destroyed.
 };*/
 
-struct Coroutine;
-Coroutine* current_coro;
-
 struct Coroutine /*: IThreadlike*/ {
-	Coroutine() = default;
-	~Coroutine() { *alive = false; }
+	Coroutine() {
+		coro = Coro_new();
+		if (coro == nullptr)
+			platform::outOfMemory();
+	}
 
-	static constexpr auto kStackSize = 256 * (1 << 10);
+	~Coroutine() { Coro_free(coro); }
 
 	void start() {
-		coro.reset(new coro_t::pull_type(boost::coroutines2::fixedsize_stack(kStackSize),
-		                                 [this](coro_t::push_type& sink) { entry(sink); }));
-		switcher(this);
+		int result = Coro_startCoro_(swapCoro(coro), coro, this, &entry);
+		if (result == ENOMEM)
+			platform::outOfMemory();
 	}
 
 	void unblock() {
 		// Coro_switchTo_( swapCoro(coro), coro );
-
-		// Copy blocked before calling send, since the call to send might destroy it.
-		auto b = blocked;
-		b.send(Void());
-	}
-
-	void waitFor(Future<Void> const& what) {
-		ASSERT(current_coro == this);
-		current_coro = nullptr;
-		(*sink)(what); // Pass control back to the switcher actor
-		ASSERT(what.isReady());
-		current_coro = this;
+		blocked.send(Void());
 	}
 
 protected:
-	ACTOR static void switcher(Coroutine* self) {
-		state std::shared_ptr<bool> alive = self->alive;
-		while (*alive && *self->coro) {
-			try {
-				wait(self->coro->get());
-			} catch (Error& e) {
-				// We just want to transfer control back to the coroutine. The coroutine will handle the error.
-			}
-			(*self->coro)(); // Transfer control to the coroutine. This call "returns" after waitFor is called.
-			wait(delay(0, g_network->getCurrentTask()));
-		}
-	}
-
-	void entry(coro_t::push_type& sink) {
-		current_coro = this;
-		this->sink = &sink;
-		run();
-	}
-
 	void block() {
 		// Coro_switchTo_( swapCoro(main_coro), main_coro );
 		blocked = Promise<Void>();
@@ -114,10 +80,16 @@ protected:
 	virtual void run() = 0;
 
 private:
-	coro_t::push_type* sink;
+	void wrapRun() {
+		run();
+		Coro_switchTo_(swapCoro(main_coro), main_coro);
+		// block();
+	}
+
+	static void entry(void* _this) { ((Coroutine*)_this)->wrapRun(); }
+
+	Coro* coro;
 	Promise<Void> blocked;
-	std::shared_ptr<bool> alive{ std::make_shared<bool>(true) };
-	std::unique_ptr<coro_t::pull_type> coro;
 };
 
 template <class Threadlike, class Mutex, bool IS_CORO>
@@ -139,11 +111,10 @@ class WorkPool final : public IThreadPool, public ReferenceCounted<WorkPool<Thre
 				delete workers[c];
 		}
 
-		ACTOR Future<Void> holdRefUntilStopped(Pool* p) {
+		Future<Void> holdRefUntilStopped(Pool* p) {
 			p->addref();
-			wait(p->allStopped.getResult());
+			co_await p->allStopped.getResult();
 			p->delref();
-			return Void();
 		}
 	};
 
@@ -202,14 +173,13 @@ class WorkPool final : public IThreadPool, public ReferenceCounted<WorkPool<Thre
 	Future<Void> m_stopOnError; // must be last, because its cancellation calls stop()!
 	Error error;
 
-	ACTOR Future<Void> stopOnError(WorkPool* w) {
+	Future<Void> stopOnError(WorkPool* w) {
 		try {
-			wait(w->getError());
+			co_await w->getError();
 			ASSERT(false);
 		} catch (Error& e) {
 			w->stop(e);
 		}
-		return Void();
 	}
 
 	void checkError() {
@@ -234,10 +204,10 @@ public:
 		pool->allStopped.add(w->stopped.getFuture());
 		startWorker(w);
 	}
-	ACTOR static void startWorker(Worker* w) {
+	static Future<Void> startWorker(Worker* w, Uncancellable = Uncancellable()) {
 		// We want to make sure that coroutines are always started after Net2::run() is called, so the main coroutine is
 		// initialized.
-		wait(delay(0, g_network->getCurrentTask()));
+		co_await delay(0, g_network->getCurrentTask());
 		w->start();
 	}
 	void post(PThreadAction action) override {
@@ -287,19 +257,44 @@ public:
 	void delref() override { ReferenceCounted<WorkPool>::delref(); }
 };
 
-typedef WorkPool<Coroutine, ThreadUnsafeSpinLock, true> CoroPool;
+using CoroPool = WorkPool<Coroutine, ThreadUnsafeSpinLock, true>;
+
+Future<Void> coroSwitcher(Future<Void> what, TaskPriority taskID, Coro* coro, Uncancellable = Uncancellable()) {
+	try {
+		// state double t = now();
+		co_await what;
+		// if (g_network->isSimulated() && g_simulator->getCurrentProcess()->rebooting && now()!=t)
+		//	TraceEvent("NonzeroWaitDuringReboot").detail("TaskID", taskID).detail("Elapsed", now()-t).backtrace("Flow");
+	} catch (Error&) {
+	}
+	co_await delay(0, taskID);
+	Coro_switchTo_(swapCoro(coro), coro);
+}
 
 void CoroThreadPool::waitFor(Future<Void> what) {
-	ASSERT(current_coro != nullptr);
+	ASSERT(current_coro != main_coro);
 	if (what.isReady())
 		return;
 	// double t = now();
-	current_coro->waitFor(what);
-	what.get(); // Throw if |what| is an error
+	coroSwitcher(what, g_network->getCurrentTask(), current_coro);
+	Coro_switchTo_(swapCoro(main_coro), main_coro);
+	// if (g_network->isSimulated() && g_simulator->getCurrentProcess()->rebooting && now()!=t)
+	//	TraceEvent("NonzeroWaitDuringReboot").detail("TaskID", currentTaskID).detail("Elapsed",
+	// now()-t).backtrace("Coro");
+	ASSERT(what.isReady());
 }
 
 // Right After INet2::run
-void CoroThreadPool::init() {}
+void CoroThreadPool::init() {
+	if (!current_coro) {
+		current_coro = main_coro = Coro_new();
+		if (main_coro == nullptr)
+			platform::outOfMemory();
+
+		Coro_initializeMainCoro(main_coro);
+		// printf("Main thread: %d bytes stack presumed available\n", Coro_bytesLeftOnStack(current_coro));
+	}
+}
 
 Reference<IThreadPool> CoroThreadPool::createThreadPool() {
 	return Reference<IThreadPool>(new CoroPool);
